% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_plots.R
\name{plot.simplexgof}
\alias{plot.simplexgof}
\title{Plots for a Bootstrap Goodness-of-Fit Test Result}
\usage{
\method{plot}{simplexgof}(
  x,
  which = c("boot", "influence"),
  ask = length(which) > 1 && dev.interactive(),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"simplexgof"} returned by
\code{\link{simplex_gof}}.}

\item{which}{Character vector indicating which plots to produce:
\code{"boot"} for the bootstrap distribution of \eqn{U_n}
(see \code{\link{plot_gof_boot}}), and/or \code{"influence"} for the
influence index plot (see \code{\link{plot_influence}}). Several can be
requested at once.}

\item{ask}{Logical; if \code{TRUE} and more than one plot is requested,
the user is asked before each new plot. Defaults to
\code{length(which) > 1 && dev.interactive()}.}

\item{...}{Further arguments passed to \code{\link{plot_gof_boot}} or
\code{\link{plot_influence}}.}
}
\value{
The object \code{x}, invisibly.
}
\description{
\code{plot} method for objects of class \code{"simplexgof"}. Produces the
bootstrap distribution of \eqn{U_n} and/or an influence index plot.
}
\examples{
data(ammonia)
X <- cbind(1, ammonia$corr_ar, ammonia$temp_agua,
           ammonia$corr_ar * ammonia$temp_agua)
Z <- cbind(1, ammonia$temp_agua,
           ammonia$corr_ar * ammonia$temp_agua)
set.seed(42)
gof <- simplex_gof(ammonia$perda, X, Z, B = 50, verbose = FALSE)
plot(gof, which = "boot")

}
\seealso{
\code{\link{simplex_gof}}, \code{\link{plot_gof_boot}},
  \code{\link{plot_influence}}
}
