% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_plots.R
\name{plot_gof_boot}
\alias{plot_gof_boot}
\title{Plot Bootstrap Distribution of the \eqn{U_n} Statistic}
\usage{
plot_gof_boot(
  x,
  col.hist = "#abd9e9",
  col.obs = "#d7191c",
  col.cv = "#2c7bb6",
  main = NULL,
  xlab = expression(U[n]^"*"),
  ylab = "Density",
  ...
)
}
\arguments{
\item{x}{A \code{"simplexgof"} object from \code{\link{simplex_gof}}.}

\item{col.hist}{Fill colour for the histogram.}

\item{col.obs}{Colour for the observed \eqn{U_n} line.}

\item{col.cv}{Colour for the critical-value lines.}

\item{main, xlab, ylab}{Plot labels.}

\item{...}{Further arguments passed to \code{\link[graphics]{hist}}.}
}
\value{
Invisibly returns the \code{"simplexgof"} object.
}
\description{
Displays the empirical bootstrap distribution of \eqn{U_n^*} together
with the observed value \eqn{U_n} and the bootstrap critical values at
each significance level, as shown in Ospina et al. (2026).
}
\examples{
\donttest{
data(ammonia)
X <- cbind(1, ammonia$corr_ar, ammonia$temp_agua,
           ammonia$corr_ar * ammonia$temp_agua)
Z <- cbind(1, ammonia$temp_agua,
           ammonia$corr_ar * ammonia$temp_agua)
set.seed(123)
gof <- simplex_gof(ammonia$perda, X, Z, B = 200, verbose = FALSE)
plot_gof_boot(gof)
}

}
\seealso{
\code{\link{simplex_gof}}, \code{\link{plot_influence}},
  \code{\link{plot.simplexgof}}
}
